# cloud-run-api/src/trendingTopicsAPI.ts
/**
 * Trending Topics API
 * AI-powered analysis of parliamentary discussion trends
 */

import express, { Request, Response } from 'express';
import NodeCache from 'node-cache';

const router = express.Router();

// Initialize services
// const db = new Firestore();
// const vertexAI = new VertexAI({ project: process.env.GOOGLE_CLOUD_PROJECT, location: 'us-central1' });

// Cache with 24-hour TTL
const cache = new NodeCache({ 
  stdTTL: 86400, // 24 hours in seconds
  checkperiod: 3600 // Check for expired keys every hour
});

// Interfaces
interface TrendingTopicSummary {
  id: string;
  title: string;
  description: string;
  trendingScore: number;
  weeklyGrowth: number;
  mentionCount: number;
  category: string;
  lastMentioned: string;
  keywords: string[];
  sources: {
    debates: number;
    questions: number;
    legislation: number;
    chatQueries: number;
  };
}

interface TrendingTopicsResponse {
  topics: TrendingTopicSummary[];
  lastUpdated: string;
  processing: boolean;
}

// GET /api/ai-insights/trending-topics
router.get('/trending-topics', async (_req: Request, res: Response) => {
  try {
    // Check cache first
    const cacheKey = 'trending-topics-list';
    let cachedData = cache.get<TrendingTopicsResponse>(cacheKey);
    
    if (cachedData) {
      console.log('Returning cached trending topics');
      return res.json(cachedData);
    }

    // Generate mock data for now (will be replaced with real processing)
    const mockTopics = await generateMockTrendingTopics();
    
    const response: TrendingTopicsResponse = {
      topics: mockTopics,
      lastUpdated: new Date().toISOString(),
      processing: false
    };

    // Cache the response
    cache.set(cacheKey, response);
    
    return res.json(response);
  } catch (error) {
    console.error('Error fetching trending topics:', error);
    return res.status(500).json({ 
      error: 'Failed to fetch trending topics',
      topics: [],
      lastUpdated: new Date().toISOString(),
      processing: false
    });
  }
});

// GET /api/ai-insights/trending-topics/:id
router.get('/trending-topics/:id', async (req: Request, res: Response) => {
  try {
    const topicId = req.params.id;
    
    // Check cache first
    const cacheKey = `trending-topic-${topicId}`;
    let cachedData = cache.get(cacheKey);
    
    if (cachedData) {
      return res.json(cachedData);
    }

    // Generate mock detailed data (will be replaced with real data)
    const topicDetail = await generateMockTopicDetail(topicId);
    
    // Cache the response
    cache.set(cacheKey, topicDetail);
    
    return res.json(topicDetail);
  } catch (error) {
    console.error(`Error fetching topic ${req.params.id}:`, error);
    return res.status(500).json({ error: 'Failed to fetch topic details' });
  }
});

// POST /api/ai-insights/process-trending
router.post('/process-trending', async (_req: Request, res: Response) => {
  try {
    // This endpoint will trigger the AI processing pipeline
    // For now, return success immediately
    
    // Clear cache to force refresh
    cache.flushAll();
    
    return res.json({ 
      message: 'Trending topics processing initiated',
      estimatedCompletion: new Date(Date.now() + 300000).toISOString() // 5 minutes
    });
  } catch (error) {
    console.error('Error processing trending topics:', error);
    return res.status(500).json({ error: 'Failed to process trending topics' });
  }
});

// Helper function to generate mock data
async function generateMockTrendingTopics(): Promise<TrendingTopicSummary[]> {
  const mockData: TrendingTopicSummary[] = [
    {
      id: 'housing-crisis-2024',
      title: 'Housing Crisis & Rental Market',
      description: 'Ongoing discussions about housing affordability, rental caps, and social housing provision across multiple parliamentary sessions.',
      trendingScore: 95.8,
      weeklyGrowth: 23.5,
      mentionCount: 147,
      category: 'housing',
      lastMentioned: new Date(Date.now() - 3600000).toISOString(), // 1 hour ago
      keywords: ['housing', 'rental', 'affordability', 'social housing', 'rental caps'],
      sources: {
        debates: 45,
        questions: 67,
        legislation: 12,
        chatQueries: 23
      }
    },
    {
      id: 'healthcare-reform-2024',
      title: 'Healthcare System Reform',
      description: 'Parliamentary focus on HSE restructuring, waiting lists, and healthcare worker conditions.',
      trendingScore: 87.2,
      weeklyGrowth: 15.8,
      mentionCount: 134,
      category: 'health',
      lastMentioned: new Date(Date.now() - 7200000).toISOString(), // 2 hours ago
      keywords: ['healthcare', 'HSE', 'waiting lists', 'medical staff', 'health reform'],
      sources: {
        debates: 38,
        questions: 52,
        legislation: 18,
        chatQueries: 26
      }
    },
    {
      id: 'climate-action-plan',
      title: 'Climate Action Implementation',
      description: 'Progress on climate targets, renewable energy initiatives, and environmental protection measures.',
      trendingScore: 76.9,
      weeklyGrowth: 31.2,
      mentionCount: 89,
      category: 'environment',
      lastMentioned: new Date(Date.now() - 14400000).toISOString(), // 4 hours ago
      keywords: ['climate', 'renewable energy', 'carbon emissions', 'environmental protection'],
      sources: {
        debates: 28,
        questions: 34,
        legislation: 15,
        chatQueries: 12
      }
    },
    {
      id: 'education-funding',
      title: 'Education Funding & Infrastructure',
      description: 'Debates on school funding, teacher shortages, and educational infrastructure development.',
      trendingScore: 69.4,
      weeklyGrowth: 8.7,
      mentionCount: 76,
      category: 'education',
      lastMentioned: new Date(Date.now() - 21600000).toISOString(), // 6 hours ago
      keywords: ['education', 'school funding', 'teachers', 'infrastructure', 'education budget'],
      sources: {
        debates: 22,
        questions: 31,
        legislation: 8,
        chatQueries: 15
      }
    },
    {
      id: 'economic-recovery',
      title: 'Post-Pandemic Economic Recovery',
      description: 'Economic policies, employment support, and business recovery measures following recent global challenges.',
      trendingScore: 64.1,
      weeklyGrowth: 12.3,
      mentionCount: 92,
      category: 'economy',
      lastMentioned: new Date(Date.now() - 28800000).toISOString(), // 8 hours ago
      keywords: ['economy', 'employment', 'business support', 'economic recovery', 'fiscal policy'],
      sources: {
        debates: 31,
        questions: 28,
        legislation: 21,
        chatQueries: 12
      }
    }
  ];

  return mockData;
}

async function generateMockTopicDetail(topicId: string) {
  // Mock detailed topic data
  return {
    topic: {
      id: topicId,
      title: 'Housing Crisis & Rental Market',
      description: 'Comprehensive analysis of housing discussions in parliament',
      // ... more detailed fields
    },
    recentMentions: [],
    timeline: [],
    relatedTopics: []
  };
}

export default router;