# scripts/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "Houses of the Oireachtas Open Data APIs",
...
    "license": {
      "name": "Oireachtas (Open Data) PSI Licence ",
      "url": "https://www.oireachtas.ie/en/open-data/license/"
    },
    "contact": {
      "name": "Oireachtas Open Data",
      "email": "open.data@oireachtas.ie",
      "url": "https://www.oireachtas.ie/en/open-data"
    }
  },
  "host": "api.oireachtas.ie",
  "basePath": "/v1",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/debates": {
      "x-swagger-router-controller": "debates",
      "get": {
        "summary": "Debates Endpoint",
        "description": "Returns list of debates filtered by the base and additional parameters.\nThe list supports paging.\n#### Mapping\n* chamber_type - debateRecord.house.chamberType\n* chamber_id - debateRecord.house.uri\n* chamber - debateRecord.house.houseCode\n* date_start (Greater Than or Equal To) - debateRecord.date\n* date_end (Less Than or Equal To) - debateRecord.date\n* skip - this will ignore the first x number of records set in the parameter\n* limit - this will only return a specific amount of records\n* member_id - debateRecord.debateSections.debateSection.speakers.speaker\n* debate_id - debateRecord.uri",
        "tags": [
          "debates"
        ],
        "operationId": "debates",
        "parameters": [
          {
            "$ref": "#/parameters/chamberType"
          },
          {
            "$ref": "#/parameters/chamberID"
          },
          {
            "$ref": "#/parameters/house"
          },
          {
            "$ref": "#/parameters/dateStart"
          },
          {
            "$ref": "#/parameters/dateEnd"
          },
          {
            "$ref": "#/parameters/skip"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/memberID"
          },
          {
            "$ref": "#/parameters/debateID"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "head": {
                  "type": "object",
                  "properties": {
                    "counts": {
                      "type": "object",
                      "properties": {
                        "debateCount": { "type": "integer" },
                        "resultCount": { "type": "integer" }
                      }
                    }
                  }
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "contextDate": { "type": "string", "format": "date" },
                      "debateRecord": {
                        "type": "object",
                        "properties": {
                          "formats": {
                            "type": "object",
                            "properties": {
                              "xml": {
                                "type": "object",
                                "properties": {
                                  "uri": { "type": "string", "format": "uri" }
                                }
                              },
                              "writtens_pdf": {
                                "type": "object",
                                "properties": {
                                  "uri": { "type": "string", "format": "uri" }
                                }
                              },
                              "pdf": {
                                "type": "object",
                                "properties": {
                                  "uri": { "type": "string", "format": "uri" }
                                }
                              }
                            }
                          },
                          "lastUpdated": { "type": "string", "format": "date-time" },
                          "debateSections": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "debateSection": {
                                  "type": "object",
                                  "properties": {
                                    "formats": {
                                      "type": "object",
                                      "properties": {
                                        "xml": {
                                          "type": "object",
                                          "properties": {
                                            "uri": { "type": "string", "format": "uri" }
                                          }
                                        },
                                        "pdf": {
                                          "type": "object",
                                          "properties": {
                                            "uri": { "type": "string", "format": "uri" }
                                          }
                                        }
                                      }
                                    },
                                    "speakers": {
                                      "type": "array",
                                      "items": {}
                                    },
                                    "debateSectionId": { "type": "string" },
                                    "containsDebate": { "type": "boolean" },
                                    "uri": { "type": "string", "format": "uri" },
                                    "parentDebateSection": {
                                      "type": "object",
                                      "properties": {
                                        "formats": {
                                          "type": "object",
                                          "properties": {
                                            "xml": {
                                              "type": "object",
                                              "properties": {
                                                "uri": { "type": "string", "format": "uri" }
                                              }
                                            },
                                            "pdf": {
                                              "type": "object",
                                              "properties": {
                                                "uri": { "type": "string", "format": "uri" }
                                              }
                                            }
                                          }
                                        },
                                        "debateSectionId": { "type": "string" },
                                        "uri": { "type": "string", "format": "uri" },
                                        "showAs": { "type": "string" }
                                      }
                                    },
                                    "bill": {
                                      "type": "object",
                                      "properties": {
                                        "event": {
                                          "type": "object",
                                          "properties": {
                                            "showAs": { "type": "string" },
                                            "houseCode": { "type": "string" },
                                            "stage": { "type": "string" },
                                            "uri": { "type": "string", "format": "uri" },
                                            "isBillStage": { "type": "boolean" }
                                          }
                                        },
                                        "uri": { "type": "string", "format": "uri" }
                                      }
                                    },
                                    "showAs": { "type": "string" },
                                    "debateType": { "type": "string" },
                                    "counts": {
                                      "type": "object",
                                      "properties": {
                                        "speechCount": { "type": "integer" },
                                        "speakerCount": { "type": "integer" }
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          },
                          "date": { "type": "string", "format": "date" },
                          "chamber": {
                            "type": "object",
                            "properties": {
                              "uri": { "type": "string", "format": "uri" },
                              "showAs": { "type": "string" }
                            }
                          },
                          "house": {
                            "type": "object",
                            "properties": {
                              "houseCode": { "type": "string" },
                              "houseNo": { "type": "string" },
                              "uri": { "type": "string", "format": "uri" },
                              "chamberType": { "type": "string" },
                              "committeeCode": { "type": "string" },
                              "showAs": { "type": "string" }
                            }
                          },
                          "debateType": { "type": "string" },
                          "uri": { "type": "string", "format": "uri" },
                          "counts": {
                            "type": "object",
                            "properties": {
                              "debateSectionCount": { "type": "integer" },
                              "questionCount": { "type": "integer" },
                              "contributorCount": { "type": "integer" },
                              "billCount": { "type": "integer" },
                              "divisionCount": { "type": "integer" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/questions": {
      "x-swagger-router-controller": "questions",
      "get": {
        "summary": "Questions Endpoint",
        "description": "Returns list of questions filtered by the base and additional parameters.\nThe list supports paging.\n#### Mapping\n* date_start (Greater Than or Equal To) - question.date\n* date_end (Less Than or Equal To) - question.date\n* skip - this will ignore the first x number of records set in the parameter\n* limit - this will only return a specific amount of records\n* qtype - question.questionType\n* member_id - question.by.uri\n* question_id - question.uri\n* question_no - question.questionNumber",
        "tags": [
          "questions"
        ],
        "operationId": "questions",
        "parameters": [
          {
            "$ref": "#/parameters/dateStart"
          },
          {
            "$ref": "#/parameters/dateEnd"
          },
          {
            "$ref": "#/parameters/skip"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/questionType"
          },
          {
            "$ref": "#/parameters/memberID"
          },
          {
            "$ref": "#/parameters/questionID"
          },
          {
            "$ref": "#/parameters/questionNo"
          },
          {
            "$ref": "#/parameters/showAnswers"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "head": {
                  "type": "object",
                  "properties": {
                    "counts": {
                      "type": "object",
                      "properties": {
                        "questionCount": {
                          "type": "integer",
                          "example": 10000
                        },
                        "resultCount": {
                          "type": "integer",
                          "example": 10000
                        }
                      }
                    }
                  }
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "question": {
                        "type": "object",
                        "properties": {
                          "answerText": { "type": "string" },
                          "questionNumber": { "type": "integer" },
                          "by": {
                            "type": "object",
                            "properties": {
                              "showAs": { "type": "string" },
                              "memberCode": { "type": "string" },
                              "uri": { "type": "string", "format": "uri" }
                            }
                          },
                          "to": {
                            "type": "object",
                            "properties": {
                              "showAs": { "type": "string" },
                              "uri": { "type": "string", "format": "uri" },
                              "roleCode": { "type": "string" },
                              "roleType": { "type": "string" }
                            }
                          },
                          "house": {
                            "type": "object",
                            "properties": {
                              "uri": { "type": "string", "format": "uri" },
                              "houseCode": { "type": "string" },
                              "houseNo": { "type": "string" },
                              "showAs": { "type": "string" },
                              "chamberType": { "type": "string" },
                              "committeeCode": { "type": "string" }
                            }
                          },
                          "uri": { "type": "string", "format": "uri" },
                          "showAs": { "type": "string" },
                          "debateSection": {
                            "type": "object",
                            "properties": {
                              "showAs": { "type": "string" },
                              "uri": { "type": "string", "format": "uri" },
                              "formats": {
                                "type": "object",
                                "properties": {
                                  "pdf": {
                                    "type": "object",
                                    "properties": {
                                      "uri": { "type": "string", "format": "uri" }
                                    }
                                  },
                                  "xml": {
                                    "type": "object",
                                    "properties": {
                                      "uri": { "type": "string", "format": "uri" }
                                    }
                                  }
                                }
                              },
                              "debateSectionId": { "type": "string" }
                            }
                          },
                          "date": { "type": "string", "format": "date" },
                          "questionType": { "type": "string" }
                        }
                      },
                      "contextDate": { "type": "string", "format": "date" }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "parameters": {
...
    "questionType": {
      "name": "qtype",
      "in": "query",
      "description": "Filter questions by oral or writtens.",
      "required": false,
      "type": "array",
      "items": {
        "type": "string",
        "enum": [
          "oral",
          "written"
        ]
      },
      "collectionFormat": "csv",
      "default": [
        "oral",
        "written"
      ],
      "maxItems": 2,
      "minItems": 0
    },
    "showAnswers": {
      "name": "show_answers",
      "in": "query",
      "description": "Add answers to the questions response",
      "required": false,
      "type": "boolean"
    },
    "memberID": {
      "name": "member_id",
      "in": "query",
      "description": "Filter by Member uri, e.g. /ie/oireachtas/member/id/Brendan-Howlin.S.1983-02-23",
      "required": false,
      "type": "string"
    },
    "debateID": {
      "name": "debate_id",
      "in": "query",
      "description": "Filter by debate uri, e.g. /akn/ie/debateRecord/dail/2024-11-07/debate/main",
      "required": false,
      "type": "string"
    },
    "chamberType": {
      "name": "chamber_type",
      "in": "query",
      "description": "Filter results by House, ie, DÃ¡il or Seanad or committees.\n",
      "required": false,
      "type": "string",
      "enum": [
        "house",
        "committee"
      ]
    },
    "house": {
      "name": "chamber",
      "in": "query",
      "description": "Filter by House name (dail or seanad).\nUsing an empty string retrieves results for both Houses.",
      "required": false,
      "type": "string",
      "default": "",
      "enum": [
        "",
        "dail",
        "seanad"
      ]
    },
    "chamberID": {
      "name": "chamber_id",
      "in": "query",
      "description": "Filter by house or committee uri, e.g. /ie/oireachtas/house/dail/32",
      "required": false,
      "type": "array",
      "collectionFormat": "csv",
      "allowEmptyValue": true,
      "default": []
    },
    "dateStart": {
      "name": "date_start",
      "in": "query",
      "description": "In the format YYYY-MM-DD (e.g. 1900-12-31)",
      "required": false,
      "type": "string",
      "format": "date"
    },
    "dateEnd": {
      "name": "date_end",
      "in": "query",
      "description": "In the format YYYY-MM-DD (e.g. 2025-12-31)",
      "required": false,
      "type": "string",
      "format": "date"
    },
...
    "skip": {
      "name": "skip",
      "in": "query",
      "description": "This allows skipping of records by a specific integer.",
      "required": false,
      "type": "integer",
      "default": 0
    },
    "limit": {
      "name": "limit",
      "in": "query",
      "description": "This allows the limiting of records to a specific integer.",
      "required": false,
      "type": "integer",
      "default": 50
    },
...
    "questionID": {
      "name": "question_id",
      "in": "query",
      "description": "Identifier for a Single Question, e.g. /ie/oireachtas/question/2025-04-10/pq_1",
      "required": false,
      "type": "string"
    },
    "questionNo": {
      "name": "question_no",
      "in": "query",
      "description": "Filter by question number",
      "required": false,
      "type": "integer"
    },
...
  }
}